refactor(abi): apply clippy suggestions for if-let chains and casts
Collapsed nested if statements using let-chains syntax, removed
unnecessary same-type casts, and simplified pattern matching.

+++ crates/frankenlibc-abi/src/stdlib_abi.rs
@@ -3977,7 +3977,7 @@ pub unsafe extern "C" fn realpath(
         )
     };
     let fd = match fd {
-        Ok(f) => f as i32,
+        Ok(f) => f,
         Err(_) => {
             unsafe { set_abi_errno(errno::ENOENT) };
             runtime_policy::observe(ApiFamily::IoFd, decision.profile, 16, true);
